Fixes for CANIO, fixes and clean up for system EEPROM settings, addition of a module to use for vehicle specific code

+++ CANIODevice.cpp
@@ -66,7 +66,7 @@ bool CANIODevice::getDigitalOutput(int which)
 	return false;
 }
 
-void CANIODevice::setAnalogOutput(int which, int16_t value)
+void CANIODevice::setAnalogOutput(int which, int value)
 {
 }
